damage log refactor

+++ src/characters/functions/dealDamage.ts
@@ -1,3 +1,5 @@
+import { AbilityLog, AddDamageBattleLog, BattleLogType } from '../../battleLog/battleLogInterfaces'
+import { addBattleLog } from '../../battleLog/functions/addBattleLog'
 import { GameState } from '../../game/GameState'
 import { DamageData, DamageTypes } from '../../items/Item'
 import { CharacterAdapter } from '../characterAdapter'
@@ -8,7 +10,8 @@ import { kill } from './kill'
 export function dealDamage(
     state: GameState,
     targetId: string,
-    damageData: DamageData
+    damageData: DamageData,
+    abilityLog: AbilityLog
 ): { state: GameState; killed: boolean; damageDone: number } {
     let killed = false
     const target = CharacterAdapter.selectEx(state.characters, targetId)
@@ -27,12 +30,20 @@ export function dealDamage(
         damageDone += damageTaken
 
         const health = Math.max(0, Math.floor(target.health - damageTaken))
-        if (health < 0.0001) {
-            state = kill(state, targetId)
-            killed = true
-        } else {
-            state = { ...state, characters: CharacterAdapter.update(state.characters, targetId, { health }) }
-        }
+
+        if (health < 0.0001) killed = true
+
+        state = { ...state, characters: CharacterAdapter.update(state.characters, targetId, { health }) }
     })
+
+    const addLog: AddDamageBattleLog = {
+        type: BattleLogType.Damage,
+        ...abilityLog,
+        damageDone,
+    }
+    state = addBattleLog(state, addLog)
+
+    if (killed) state = kill(state, targetId)
+
     return { state, killed, damageDone }
 }
